Preliminary setup, remove blst code, refactor `cryoto` into own module
