+++ soroban-env-host/src/host.rs
@@ -24,7 +24,6 @@ use crate::{
 
 mod comparison;
 mod conversion;
-pub(crate) mod crypto;
 mod data_helper;
 mod declared_size;
 pub(crate) mod error;
@@ -38,7 +37,7 @@ pub(crate) mod metered_map;
 pub(crate) mod metered_vector;
 pub(crate) mod metered_xdr;
 mod num;
-mod prng;
+pub(crate) mod prng;
 pub(crate) mod trace;
 mod validity;
